separate config and simdata tests
