+++ src/simoc_abm/agents/lamp.py
